fix(examples): make XYsafe return a signed type (#433)
XYsafe() returns -1 to steer out-of-bounds coordinates onto the safety
pixel, but was declared uint16_t, so -1 became 65535.

That only happens to work where int is 16 bits wide. On AVR, adding 65535
to a 16-bit index wraps to "subtract 1" and leds[XYsafe(...)] lands on
leds_plus_safety_pixel[0] as intended. On any 32-bit board the value
promotes to the int 65535 with no wraparound, so the same expression
indexes ~65 KB past a matrix that is typically a few hundred bytes.

Declaring it `int` makes the -1 an actual -1 on every platform. `int`
rather than int16_t so 32-bit boards keep a type wide enough for any
matrix they can hold; on AVR it is 16 bits, which is the width that
already worked there.

This also fixes the debugging trap in the original report: printing the
result showed 65535, which reads as a bogus index rather than as the
out-of-bounds sentinel actually being returned.

Diagnosed by @eltang in the issue thread; @kriegsman reopened it agreeing
32-bit platforms likely had a real bug here.

Refs #433

+++ examples/XYMatrix/XYMatrix.ino
@@ -172,7 +172,14 @@ uint16_t XY( uint8_t x, uint8_t y)
 CRGB leds_plus_safety_pixel[ NUM_LEDS + 1];
 CRGB* const leds( leds_plus_safety_pixel + 1);
 
-uint16_t XYsafe( uint8_t x, uint8_t y)
+// NOTE: the return type must be SIGNED. Returning -1 from a uint16_t makes the
+// value 65535, which only behaves like -1 where int is 16 bits wide (AVR). On a
+// 32-bit board it promotes to the int 65535, so leds[XYsafe(...)] indexes far
+// past the array instead of landing on the safety pixel.
+//
+// `int` rather than `int16_t` so that on 32-bit boards the type stays wide
+// enough for any matrix a board that size can actually hold.
+int XYsafe( uint8_t x, uint8_t y)
 {
   if( x >= kMatrixWidth) return -1;
   if( y >= kMatrixHeight) return -1;
